♻️(ats-domain) move etape_recrutement in shared
shared by recrutement and candidature bounded context
