+++ src/web/domain/shared/exceptions/erreur_shared.py
